requirements.txt
qrcode[pil]>=7.3
Pillow>=9.0
Jinja2>=3.0
